requirements.txt
rasterio
numpy
pillow